backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
reportlab
